Split ITs from launcher
