+++ nexus/nexus-test/nexus-testsuite-support/src/test/resources/logback-test.xml
